Update gatsby peer dependency
Gatsby 4 was released on Oct 25, 2021. Since then projects using this plugin got a warning message.
This means projects may already use the latest gatsby version with this plugin. As there are currently no 
open issues for known compatibility errors open it should be safe to update the peer dependency.

Fixes #25